Refactor ObjectDetection to domain layer
